+++ Changes.md
@@ -24,6 +24,7 @@ Fixes
 - Cycles :
   - Fixed crash rendering a scene containing a CoordinateSystem (#4865).
   - Fixed crash rendering a PointPrimitive with `N` primitive variable (#4876).
+  - Fixed handling of `Varying` primitive variable interpolation (#4849).
 - PathFilter : Fixed error when selecting a path element from a promoted `PathFilter.paths` plug (introduced in 0.61.13.0).
 
 API